backend/requirements.txt
Flask==2.2.3
flask_cors==4.0.0
matplotlib==3.5.2
munkres==1.1.4
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==10.2.0
requests==2.28.1
scipy==1.12.0
skimage==0.0
torch==2.2.0
torchvision==0.17.0
